+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Usage:
 * Fix #2301: Add compatibility with SemVer versions
 * Fix #2302 Bump Kubernetes Client version to 6.8.0
 * Fix #2324: Update SpringBootConfigurationHelper for Spring Boot 3.x
+* Fix #2336: Add new `spring` build strategy which delegates to spring boot's build image
 * Fix #2350: Enrichers with NAME configuration override fragments with default names
 
 ### 1.13.1 (2023-06-16)